requirements.txt
openai
pandas
python-dotenv
beautifulsoup4